apps/modelApi/requirements.txt
pickle5==0.0.11
pathlib==1.0.1
pandas==1.5.3
scikit-learn==1.2.2
python-dotenv==1.0.1